[KFI] LucQuery and LucObject moved into the Lucene29 namespace.
