+++ src/ContentRepository/Linq/LinqChildrenEnumerator.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using SenseNet.Search;
+using SenseNet.Search.Lucene29;
 
 namespace SenseNet.ContentRepository.Linq
 {
